feat: move all models to models folder
