+++ packages/api/nest-cli.json
@@ -1,4 +1,5 @@
 {
   "collection": "@nestjs/schematics",
-  "sourceRoot": "src"
+  "sourceRoot": "src/modules",
+  "entryFile": "src/main"
 }
