+++ packages/api/src/interfaces/cookie-types.d.ts
@@ -1,7 +1,7 @@
 import { Request } from '@nestjs/common';
 import { CookieOptions } from 'express';
 
-import { UserEntity } from '../database/entities/user.entity';
+import { UserEntity } from '../modules/database/entities/user.entity';
 
 type Modify<T, R> = Omit<T, keyof R> & R;
 export interface LoginRequest
