+++ packages/api/src/modules/users/users.service.ts
@@ -1,8 +1,8 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { Connection } from 'typeorm';
 
+import { GetList } from '../../interfaces/react-admin-types';
 import { UserEntity } from '../database/entities/user.entity';
-import { GetList } from '../interfaces/react-admin-types';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
