+++ packages/api/test/utils/clearTestDatabase.ts
@@ -1,7 +1,7 @@
 import { Test } from '@nestjs/testing';
 import { Connection } from 'typeorm';
 
-import { DatabaseModule } from '../../src/database/database.module';
+import { DatabaseModule } from '../../src/modules/database/database.module';
 
 export const clearTestDatabase = async (): Promise<void> => {
   const app = await Test.createTestingModule({
